requirements.txt
asgiref==3.4.1
attrs==21.2.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.0
cssselect2==0.4.1
defusedxml==0.7.1
Django==3.2.9
django-debug-toolbar==3.2.2
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-nested-routers==0.93.4
drf-spectacular==0.21.0
easy-thumbnails==2.8
google-auth==2.3.3
idna==3.3
importlib-resources==5.4.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jsonschema==4.2.1
lxml==4.6.4
MarkupSafe==2.0.1
mysqlclient==2.1.0
oauthlib==3.1.1
Pillow==8.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyrsistent==0.18.0
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
reportlab==3.6.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
svglib==1.1.0
tinycss2==1.1.1
uritemplate==4.1.1
urllib3==1.26.7
webencodings==0.5.1
zipp==3.6.0
